Extended track
Extended track so it isn't annoying after 2nd track, well it's getting annoying for me, because of my voice, so the track is now 3 minutes long and the vocals kick in after one minute and play until we go back to main menu. In main menu the track resets and starts playing again until you decide to turn the music off.... you can't obviously xD.
Anyways gotta add a slider or a button for muting music.